Added Multithreading proejct, but no work done yet.
Check-In is mostly for more refactoring around the Maze Game.  Refactored the Collision so that the work in the HandleCollision method is sent off to a helper class.  More of an organizational change then anything.

Thought about adding collison to the individual PlacableActor's, but didn't like the idea of Actors dealing with something that the GameState should handle.