Doing a good bit too the program.
Server now acts as a log instead of a chat member.  Program now expects multiple clients to communicate with each other.

Adding in some formatting for the inputting and outputting of messages on the client side.

Kinda messy.  Might roll through it later this week and clean up some if I have time between the other challenges.